[skstore:ts] Reorganize and cleaning
